[Refactor] #172 - MyPage 화면에서 장르·키워드·서재 서브뷰를 공용 컴포넌트로 분리
